[Code] 서버/클라이언트 레이아웃 모듈 분리
